Send Attendance Report Email
Type(s): UPDATE

Issue(s):
- ClickUp: 2unh7m2

+++ app/Http/Controllers/B2b/AttendanceController.php
@@ -166,7 +166,7 @@ class AttendanceController extends Controller
 
         $business_members = $business->getAllBusinessMemberExceptInvited();
 
-        if ($request->has('department_id')) {
+        if ($request->has('department_id') && $request->department_id != 'null') {
             $business_members = $business_members->whereHas('role', function ($q) use ($request) {
                 $q->whereHas('businessDepartment', function ($q) use ($request) {
                     $q->where('business_departments.id', $request->department_id);
@@ -174,7 +174,7 @@ class AttendanceController extends Controller
             });
         }
 
-        if($request->has('status')) {
+        if($request->has('status') && $request->status != 'null') {
             $business_members = $business_members->where('status', $request->status);
         }
         $business_members = $business_members->get();
@@ -234,7 +234,7 @@ class AttendanceController extends Controller
 
         $all_employee_attendance = $this->filterInactiveCoWorkersWithData($all_employee_attendance);
 
-        if ($request->has('search')) $all_employee_attendance = $this->searchWithEmployeeName($all_employee_attendance, $request);
+        if ($request->has('search') && $request->search != 'null') $all_employee_attendance = $this->searchWithEmployeeName($all_employee_attendance, $request);
         if ($request->has('sort_on_absent')) $all_employee_attendance = $this->attendanceSortOnAbsent($all_employee_attendance, $request->sort_on_absent);
         if ($request->has('sort_on_present')) $all_employee_attendance = $this->attendanceSortOnPresent($all_employee_attendance, $request->sort_on_present);
         if ($request->has('sort_on_leave')) $all_employee_attendance = $this->attendanceSortOnLeave($all_employee_attendance, $request->sort_on_leave);
@@ -242,10 +242,15 @@ class AttendanceController extends Controller
         if ($request->has('sort_on_overtime')) $all_employee_attendance = $this->attendanceCustomSortOnOvertime($all_employee_attendance, $request->sort_on_overtime);
 
         if ($request->file == 'excel') {
-            $monthly_excel->setMonthlyData($all_employee_attendance->toArray())->setStartDate($request->start_date)->setEndDate($request->end_date)->get();
-            $file_path = storage_path('exports').'/Custom_attendance_report.xls';
-            (new SendMonthlyAttendanceReportEmail($file_path, $request->business_member, $start_date, $end_date))->handle();
-            unlink($file_path);
+            if ($request->business_member->member->profile->email == 'raz@sheba.xyz') {
+                $monthly_excel->setMonthlyData($all_employee_attendance->toArray())->setStartDate($request->start_date)->setEndDate($request->end_date)->save();
+                $file_path = storage_path('exports') . '/Custom_attendance_report.xls';
+                dispatch(new SendMonthlyAttendanceReportEmail($file_path, $request->business_member, $start_date, $end_date));
+                unlink($file_path);
+                return api_response($request, null, 200, ['message' => "Your report will be sent to your email. Please check a few moments later."]);
+            } else {
+                $monthly_excel->setMonthlyData($all_employee_attendance->toArray())->setStartDate($request->start_date)->setEndDate($request->end_date)->get();
+            }
         }
 
         return api_response($request, $all_employee_attendance, 200, ['all_employee_attendance' => $all_employee_attendance, 'total_members' => $total_business_members_count]);
